software/requirements.txt
CherryPy==3.3.0
Jinja2==2.7.3
MarkupSafe==0.23
WebOb==1.4
coverage==3.7.1
euclid==0.01
fanstatic==0.16
flake8==2.2.2
js.jquery==1.9.1
mccabe==0.2.1
nose==1.3.3
pep8==1.5.7
pyflakes==0.8.1
pyserial==2.7
web.py==0.37
wsgiref==0.1.2
